api_server_edrian/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
pydantic==2.5.0
requests==2.31.0
PyJWT==2.8.0
six
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
typing